:art: Break apart PHP megaclass into multiple classes
